refactor($GitHub): check outdated Python packages in GitHub Actions
This commit introduces a new job in the GitHub Actions workflow that
checks for outdated Python packages. The job runs the command `pipenv
run pip list --outdated` and stores its output. If the output is empty,
it prints "Congrats! Everything is up-to-date". If it's not empty, it
prints "Attention! Outdated dependencies detected" and the list of
outdated packages, then exits with a status code of 1, causing the job
to fail. This way, we can be alerted of outdated dependencies by a
failing GitHub Actions job.